requirements.txt
streamlit>=1.30.0
yfinance>=0.2.30
tensorflow>=2.15.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
anthropic>=0.40.0
xgboost>=2.0.0
joblib>=1.3.0
scipy>=1.11.0
statsmodels>=0.14.0
